+++ src/termout.c
@@ -3097,7 +3097,7 @@ do_csi(uchar c)
     when 'm':        /* SGR: set graphics rendition */
       do_sgr();
 #if 0
-    /* added in 3.6.2 (#1171), withdrawn in 3.6.3 (conflict with XTQMODKEYS */
+    /* added in 3.6.2 (#1171), withdrawn in 3.6.3 (conflict with XTQMODKEYS) */
     when CPAIR('?', 'm'):  /* DEC private SGR (EK-PPLV2-PM-B01) */
       switch (arg0) {
         when 4: term.curs.attr.attr &= ~ATTR_SUBSCR;
@@ -4373,7 +4373,7 @@ do_cmd(void)
         term.curs.attr.link = -1;
     }
     when 60: {  // xterm XTQALLOWED: query allowed runtime features
-      child_printf("\e]62;%s%s%s%s%s", 
+      child_printf("\e]60;%s%s%s%s%s", 
         // check foreground/background colour setting as an approximation
         contains(cfg.suppress_osc, 10) || contains(cfg.suppress_osc, 11) 
           ? "" : ",allowColorOps",
@@ -4383,15 +4383,15 @@ do_cmd(void)
         //allowWindowOps
         osc_fini());
     }
-    when 61: {  // xterm XTQALLOWED: query disallowed runtime subfeatures
+    when 61: {  // xterm XTQDISALLOWED: query disallowed runtime subfeatures
       if (!strcasecmp(s, "allowColorOps"))
-        child_printf("\e]62%s%s",
+        child_printf("\e]61%s%s",
           contains(cfg.suppress_osc, 4) ? ";GetAnsiColor,SetAnsiColor" : "",
           osc_fini());
       else if (!strcasecmp(s, "allowFontOps"))
-        child_printf("\e]62;GetFont,SetFont%s", osc_fini());
+        child_printf("\e]61;GetFont,SetFont%s", osc_fini());
       else if (!strcasecmp(s, "allowMouseOps"))
-        child_printf("\e]62;VT200Hilite%s%s%s%s%s%s%s%s%s%s%s%s%s%s",
+        child_printf("\e]61;VT200Hilite%s%s%s%s%s%s%s%s%s%s%s%s%s%s",
           strstr(cfg.suppress_wheel, "scrollwin") ? ",Scrollback" : "",
           strstr(cfg.suppress_wheel, "zoom") ? ",ZoomMouse" : "",
           strstr(cfg.suppress_wheel, "report") ? ",WheelEvent" : "",
@@ -4410,9 +4410,9 @@ do_cmd(void)
           contains(cfg.suppress_dec, 1004) ? ",FocusEvent" : "",
           osc_fini());
       else if (!strcasecmp(s, "allowTitleOps"))
-        child_printf("\e]62%s", osc_fini());
+        child_printf("\e]61%s", osc_fini());
       else if (!strcasecmp(s, "allowWindowOps"))
-        child_printf("\e]62;SetChecksum,SetXprop,GetIconTitle,GetWinTitle%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s",
+        child_printf("\e]61;SetChecksum,SetXprop,GetIconTitle,GetWinTitle%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s%s",
           contains(cfg.suppress_osc, 52) ? ",GetSelection" : "",
           (contains(cfg.suppress_osc, 52) || !cfg.allow_set_selection)
             ? "SetSelection" : "",
@@ -4436,9 +4436,9 @@ do_cmd(void)
           contains(cfg.suppress_win, 4) ? ",SetWinSizePixels" : "",
           osc_fini());
       else if (!strcasecmp(s, "allowTcapOps"))
-        child_printf("\e]62;SetTcap,GetTcap%s", osc_fini());
+        child_printf("\e]61;SetTcap,GetTcap%s", osc_fini());
       else if (!strcasecmp(s, "allowPasteControls"))
-        child_printf("\e]62;%s%s", cfg.filter_paste, osc_fini());
+        child_printf("\e]61;%s%s", cfg.filter_paste, osc_fini());
     }
     when 1337: {  // iTerm2 image protocol
                   // https://www.iterm2.com/documentation-images.html